Fixed a bug in initializing the default competition
The test data inserter initializes the default competition upon inserting test data. However, now this happens only upon the first start. In subsequent starts data is not inserted anymore.

Changed the way the default competition is initialized in cases when data is not inserted. This also removed the need for the AutoInsert fake entity.